+++ projects/customise/src/app/app-routing.module.ts
@@ -14,13 +14,17 @@ const routes: Routes = [
     loadChildren: (): any => import('projects/customise/src/app/writer/writer.module').then(m => m.WriterModule),
   },
   {
-    path:'customise-body-server-log',
+    path: 'customise-body-server-log',
     loadChildren: (): any => import('projects/customise/src/app/customise-body-server-log/customise-body-server-log.module').then(m => m.CustomiseBodyServerLogModule),
+  },
+  {
+    path: 'auth-token',
+    loadChildren: (): any => import('projects/customise/src/app/auth-token/auth-token.module').then(m => m.AuthTokenModule),
   },
   {
     path: 'main',
     component: MainComponent,
-  }
+  },
 ];
 
 @NgModule({